fix a bug in bookpuns
The logic of the bookpuns function relied on there being only one colon
per line separating the title and the author.
Fun fact, there were two lines with two colons.
If ever a line was added that omitted the colon, that would also break it.

At first I thought to write more careful code for getting the title and
author, but then I realised: why are we reformatting the line every time?
Just write "<title> by <author>" in the data file, boom, solved.